Light organising

+++ TombEngine/Objects/TR5/Object/tr5_pushableblock.h
@@ -14,8 +14,9 @@ namespace TEN::Entities::Generic
 	void PushableBlockControl(short itemNumber);
 	void PushableBlockCollision(short itemNumber, ItemInfo* laraItem, CollisionInfo* coll);
 	bool TestBlockMovable(ItemInfo* item, int blockHeight);
-	bool TestBlockPush(ItemInfo* item, int blockHeight, unsigned short quadrant);
-	bool TestBlockPull(ItemInfo* item, int blockHeight, short quadrant);
+	bool TestBlockPush(ItemInfo* item, int blockHeight, int quadrant);
+	bool TestBlockPull(ItemInfo* item, int blockHeight, int quadrant);
+
 	void MoveStackXZ(short itemNumber);
 	void MoveStackY(short itemNumber, int y);
 	void RemoveBridgeStack(short itemNumber);
@@ -24,8 +25,10 @@ namespace TEN::Entities::Generic
 	int FindStack(short itemNumber);
 	int GetStackHeight(ItemInfo* item);
 	bool CheckStackLimit(ItemInfo* item);
+
 	void PushLoop(ItemInfo* item);
 	void PushEnd(ItemInfo* item);
+
 	std::optional<int> PushableBlockFloor(short itemNumber, int x, int y, int z);
 	std::optional<int> PushableBlockCeiling(short itemNumber, int x, int y, int z);
 	int PushableBlockFloorBorder(short itemNumber);
